Fix SQL formatting issue.
We need to end formatted SQL lines with a space because the Tools delivered formatter first collapses things to a single line and then does its formatting. we the format their formatted text. If we don't ensure lines end with a space this "collapse to a single line" ends up sticking unrelated tokens together.

Fixes #45